feat: refactor all state management to use redux
