+++ frontend/src/components/Buttons/OperatorButton.tsx
@@ -1,26 +1,27 @@
-import React, { useContext } from "react";
-import { ActionType } from "../../contexts/Actions";
+import React from "react";
+import { useAppDispatch, useAppSelector } from "../../store/hooks";
 import {
-  CalculatorContext,
-  CalculatorDispatchContext,
-} from "../../contexts/CalculatorContext";
+  press_operator,
+  selectLeftOperand,
+  selectOperator,
+  selectRightOperand,
+} from "../../store/slices/calculatorSlice";
 
 export default function OperatorButton({ label }: { label: string }) {
-  const state = useContext(CalculatorContext);
-  const dispatch = useContext(CalculatorDispatchContext);
+  const leftOperand = useAppSelector(selectLeftOperand);
+  const rightOperand = useAppSelector(selectRightOperand);
+  const operator = useAppSelector(selectOperator);
+  const dispatch = useAppDispatch();
 
   function handleClick(e: any) {
     e.preventDefault();
-    dispatch({
-      type: ActionType.OPERATOR_PRESSED,
-      payload: label,
-    });
+    dispatch(press_operator(label));
   }
 
   if (
-    state.leftOperand.length !== 0 &&
-    state.operator === label &&
-    state.rightOperand.length === 0
+    leftOperand.length !== 0 &&
+    operator === label &&
+    rightOperand.length === 0
   ) {
     return (
       <button
