+++ frontend/src/components/KeyboardInput/KeyboardInput.tsx
@@ -1,56 +1,43 @@
-import React, { useContext, useEffect } from "react";
+import React, { useEffect } from "react";
 import {
   NumberButtonLabels,
   OperatorButtonLabels,
 } from "../../constants/Constants";
-import { ActionType } from "../../contexts/Actions";
-import { CalculatorDispatchContext } from "../../contexts/CalculatorContext";
+import { useAppDispatch } from "../../store/hooks";
+import {
+  press_clear,
+  press_decimal,
+  press_equals,
+  press_number,
+  press_operator,
+  press_percent,
+} from "../../store/slices/calculatorSlice";
 
 // The KeyboardInput component does not render anything to the DOM but handles keyboard inputs.
 export default function KeyboardInput() {
-  const dispatch = useContext(CalculatorDispatchContext);
+  const dispatch = useAppDispatch();
 
   const handleKeyUp = (event: any) => {
     // Convert the keypress to lowercase to handle things like "c", "x", "backspace", "delete" reliably
     const key: string = event.key.toLowerCase();
 
     if (NumberButtonLabels.includes(key)) {
-      dispatch({
-        type: ActionType.NUMBER_PRESSED,
-        payload: key,
-      });
+      dispatch(press_number(key));
     } else if (OperatorButtonLabels.includes(key)) {
       // Handles "+", "-"
-      dispatch({
-        type: ActionType.OPERATOR_PRESSED,
-        payload: key,
-      });
+      dispatch(press_operator(key));
     } else if (key === "/") {
-      dispatch({
-        type: ActionType.OPERATOR_PRESSED,
-        payload: "÷",
-      });
+      dispatch(press_operator("÷"));
     } else if (key === "x" || key === "*") {
-      dispatch({
-        type: ActionType.OPERATOR_PRESSED,
-        payload: "x",
-      });
+      dispatch(press_operator("x"));
     } else if (key === ".") {
-      dispatch({
-        type: ActionType.DECIMAL_PRESSED,
-      });
+      dispatch(press_decimal());
     } else if (key === "c" || key === "backspace" || key === "delete") {
-      dispatch({
-        type: ActionType.CLEAR_PRESSED,
-      });
+      dispatch(press_clear());
     } else if (key === "=" || key === "enter") {
-      dispatch({
-        type: ActionType.EQUALS_PRESSED,
-      });
+      dispatch(press_equals());
     } else if (key === "%") {
-      dispatch({
-        type: ActionType.PERCENT_PRESSED,
-      });
+      dispatch(press_percent());
     }
     return;
   };
