--- frontend/src/contexts/Actions.ts
@@ -1,9 +1,0 @@
-export enum ActionType {
-  "NUMBER_PRESSED",
-  "OPERATOR_PRESSED",
-  "EQUALS_PRESSED",
-  "CLEAR_PRESSED",
-  "DECIMAL_PRESSED",
-  "PERCENT_PRESSED",
-  "PLUS_MINUS_PRESSED",
-}
