--- frontend/src/contexts/CalculatorContext.ts
@@ -1,7 +1,0 @@
-import { createContext, Dispatch } from "react";
-import { IState } from "./types";
-
-export const CalculatorContext = createContext<IState>({} as IState);
-export const CalculatorDispatchContext = createContext<Dispatch<any>>(
-  {} as Dispatch<any>
-);
