--- frontend/src/contexts/types.ts
@@ -1,6 +1,0 @@
-export interface IState {
-  leftOperand: string;
-  rightOperand: string;
-  operator: string;
-  result: string;
-}
