+++ frontend/src/store/store.ts
@@ -1,0 +1,13 @@
+import { configureStore } from "@reduxjs/toolkit";
+import calculatorReducer from "../store/slices/calculatorSlice";
+
+export const store = configureStore({
+    reducer: {
+        calculator: calculatorReducer,
+    }
+});
+
+// Infer the `RootState` and `AppDispatch` types from the store itself
+export type RootState = ReturnType<typeof store.getState>
+// Inferred type: {posts: PostsState, comments: CommentsState, users: UsersState}
+export type AppDispatch = typeof store.dispatch
